Add files via upload
The user is required to edit the folder names that need to be created on running this program. Once the names are set, the python program will automatically create the folders in the Documents folder. This code is set to run in Linux OS by default. If running in Windows, the user needs to comment out the Linux path and remove the comment quotes from the Windows path.